Move to useValidatePassword hook instead of redux action
