+++ frontend/src/metabase/auth/actions.ts
@@ -1,5 +1,4 @@
 import { type UnknownAction, createAction } from "@reduxjs/toolkit";
-import { getIn } from "icepick";
 import { push } from "react-router-redux";
 
 import { loadLocalization } from "metabase/api/localization";
@@ -9,11 +8,10 @@ import { clearCurrentUser, refreshCurrentUser } from "metabase/redux/user";
 import { createAsyncThunk } from "metabase/redux/utils";
 import { getSetting } from "metabase/selectors/settings";
 import { getUser } from "metabase/selectors/user";
-import { SessionApi, UtilApi } from "metabase/services";
+import { SessionApi } from "metabase/services";
 import * as Urls from "metabase/urls";
 import { isSmallScreen, reload } from "metabase/utils/dom";
 import { isResourceNotFoundError } from "metabase/utils/errors";
-import { passwordComplexityDescription } from "metabase/utils/password";
 
 import type { LoginData } from "./types";
 
@@ -165,19 +163,6 @@ export const resetPassword = createAsyncThunk(
     }
   },
 );
-
-export const validatePassword = async (password: string) => {
-  const error = passwordComplexityDescription(password);
-  if (error) {
-    return error;
-  }
-
-  try {
-    await UtilApi.password_check({ password });
-  } catch (error) {
-    return getIn(error, ["data", "errors", "password"]);
-  }
-};
 
 const initiateSLO = async () => {
   try {
