+++ frontend/src/metabase/auth/components/ResetPassword/ResetPassword.tsx
@@ -6,10 +6,11 @@ import { t } from "ttag";
 import { useGetPasswordResetTokenStatusQuery } from "metabase/api";
 import { Button } from "metabase/common/components/Button";
 import { Link } from "metabase/common/components/Link";
+import { useValidatePassword } from "metabase/common/hooks";
 import { useToast } from "metabase/common/hooks/use-toast";
 import { useDispatch } from "metabase/redux";
 
-import { resetPassword, validatePassword } from "../../actions";
+import { resetPassword } from "../../actions";
 import type { ResetPasswordData } from "../../types";
 import { AuthLayout } from "../AuthLayout";
 import { ResetPasswordForm } from "../ResetPasswordForm";
@@ -35,6 +36,7 @@ export const ResetPassword = ({
   const email = location?.query?.email;
   const dispatch = useDispatch();
   const [sendToast] = useToast();
+  const validatePassword = useValidatePassword();
   const { data: status, isLoading } =
     useGetPasswordResetTokenStatusQuery(token);
