+++ frontend/src/metabase/common/hooks/index.ts
@@ -12,6 +12,7 @@ export * from "./use-temp-storage";
 export * from "./use-temporary-state";
 export * from "./use-toast";
 export * from "./use-url-with-utm";
+export * from "./use-validate-password";
 export { useStoreUrl } from "./use-store-url/use-store-url";
 export * from "./use-capture-event";
 export * from "./use-progressive-loader";
